test: split test in multiple files
